Multiple fixes to SynergyKM for 10.7 support:
- Fixed 64-bit issues and enabled garbage collection support
- Fixed project file to include proper build products and resources
- Modified calls to NSDistributedNotificationCenter to use the option
  'NSNotificationDeliverImmediately' so that the daemon will receive
  notifications properly.
- Fixed a few warnings

+++ Components/Prefpane/Controllers/SPServerTabController.h
@@ -67,7 +67,7 @@
 
     NSView * superview;
 
-    unsigned int selectedIndex;
+	NSInteger selectedIndex;
     
     NSPoint * cachedPositions;
 }
